pyproject.toml
[tool.poetry]
name = "assignment-1"
version = "0.1.0"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
azure-storage-blob = "^12.23.0"
streamlit-option-menu = "^0.3.13"
openai = "^1.47.0"
requests = "^2.32.3"
pymssql = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


